issue #7 - deleting user working

+++ Hades/Hades/ClientApp/src/app/webPages/groupRoom/components/navbar/navbar.component.ts
@@ -1,6 +1,9 @@
 import { Component, OnInit } from '@angular/core';
 import { CookieService } from 'ngx-cookie-service';
 import { Router } from '@angular/router';
+import { BackendcallService } from 'src/app/core/httpsCalls/backendcall.service';
+
+import { ToastrServiceService } from 'src/app/core/notifications/toastr-service.service';
 
 @Component({
   selector: 'app-navbar',
@@ -18,7 +21,9 @@ export class NavbarComponent implements OnInit {
 
   constructor(
     private cookie: CookieService,
-    private router: Router
+    private router: Router,
+    private backendcall: BackendcallService,
+    private toust: ToastrServiceService
   ) { }
 
 
@@ -38,8 +43,21 @@ export class NavbarComponent implements OnInit {
   }
 
   closeGroup(){
-    this.cookie.deleteAll();
-    this.router.navigate(["/main"]);
+    this.backendcall.deteleUser(this.cookie.get("userId")).subscribe( res => {
+      console.log(res);
+      if(res.result){
+        this.toust.showSuccess("Success", res.resultText);
+        this.cookie.deleteAll();
+        this.router.navigate(["/main"]);
+      }else{
+        this.toust.showError("Error",res.resultText);
+      }
+
+    }, error => {
+      console.log(error);
+      this.toust.showError("An error!", "Be kind and try again later!");
+    });
+
   }
 
 }
